backend/requirements.txt
flask
python-dotenv
flask-restful
flask-sqlalchemy
flask_migrate
flask_jwt_extended
flask-mail
faker


